Add searchbar to container in corresponding file
