+++ index.js
@@ -1,12 +1,7 @@
-import { createCharacterCard } from "./components/card/card.js";
+import { createCharacterCards } from "./components/card/card.js";
 import { createSearchBar } from "./components/search-bar/search-bar.js";
 import { createPagination } from "./components/nav-pagination/nav-pagination.js";
 import { changePaginationContent } from "../components/nav-pagination/nav-pagination.js";
-
-const cardContainer = document.querySelector('[data-js="card-container"]');
-const searchBarContainer = document.querySelector(
-  '[data-js="search-bar-container"]'
-);
 
 // States
 let page = 1;
@@ -15,8 +10,7 @@ let fetchurl = "";
 let maxPage = 1;
 
 // Creating and appending the search bar
-searchBarContainer.append(createSearchBar());
-
+createSearchBar();
 // Create DOM Elemnts for the navigation bar
 await createPagination();
 // Initial fill of the page with the cards
@@ -28,23 +22,11 @@ export async function updateCharacterCards(newPage = 1, newSearchQuery = "") {
     if (newSearchQuery) searchQuery = newSearchQuery;
     fetchurl = setFetchURL(page, searchQuery);
 
-    cardContainer.innerHTML = "";
     const charactersData = await fetch(fetchurl);
     const charactersDataJson = await charactersData.json();
     const characters = charactersDataJson.results;
-    if (characters) {
-      characters.forEach((character) => {
-        cardContainer.append(
-          createCharacterCard(
-            character.name,
-            character.status,
-            character.type,
-            character.episode.length,
-            character.image
-          )
-        );
-      });
-    }
+    createCharacterCards(characters);
+
     maxPage = charactersDataJson.info.pages;
     changePaginationContent(maxPage, page);
   } catch (error) {
